Apisimilar/word_similarity_service/requirements.txt
fastapi
uvicorn
fasttext-wheel
pydantic
httpx
ratelimit
backoff
